警告問題箇所以外修正

+++ Engine/3d/FbxLoader.cpp
@@ -254,7 +254,7 @@ void FbxLoader::ParseMeshFaces(FbxModel* model, FbxMesh* fbxMesh)
             //3頂点目までなら
             if (j < 3) {
                 //1頂点追加し、他の2点と三角形を構築する
-                indices.push_back(index);
+                indices.push_back((unsigned short)index);
             }
             //4頂点目
             else {
@@ -262,9 +262,9 @@ void FbxLoader::ParseMeshFaces(FbxModel* model, FbxMesh* fbxMesh)
                 int32_t index2 = indices[indices.size() - 1];
                 int32_t index3 = index;
                 int32_t index0 = indices[indices.size() - 3];
-                indices.push_back(index2);
-                indices.push_back(index3);
-                indices.push_back(index0);
+                indices.push_back((unsigned short)index2);
+                indices.push_back((unsigned short)index3);
+                indices.push_back((unsigned short)index0);
             }
         }
     }
